requirements.txt
gunicorn==19.9.0
pugsql==0.1.9
varyaml==0.0.4
flask==1.0.3
bcrypt==3.1.6
statsd==3.3.0
redis==3.2.1
psycopg2-binary==2.8.2
uuid==1.30
